Misc fixes (#19998)
* Misc fixes

- Change disp-name on vmk workload to host's disp name
- handle host disp name property when received w/o config
- add checking in migration start so that interfacce cannot be erroneously changed via REST
- Changed Assert -> AssertEquals
- Moved DeleteHost() only where vcHub is destroyed due to delete of orchConfig (changed the flag name to reflect it)
- improved test coverage

* - fixed the duplicate vchub.Destroy() in test cases
- fixed the name label code, had incorrect lines copied and used addNamespaceLabel() rather than
	addNameLabel() function

+++ venice/ctrler/orchhub/orchestrators/vchub/vchub.go
@@ -149,22 +149,22 @@ func (v *VCHub) createProbe(config *orchestration.Orchestrator) {
 }
 
 // Destroy tears down VCHub instance
-func (v *VCHub) Destroy(cleanRemote bool) {
+func (v *VCHub) Destroy(delete bool) {
 	// Teardown probe and store
 	v.Log.Infof("Destroying VCHub....")
 
 	// Clearing probe/session state after all routines finish
 	// so that a thread in the middle of writing doesn't get a nil client
-	if cleanRemote {
+	if delete {
 		v.Log.Infof("Cleaning up state on VCenter.")
 		v.deleteAllDVS()
+		v.DeleteHosts()
 	}
 	v.cancel()
 	v.Wg.Wait()
 
 	v.probe.ClearState()
 
-	v.DeleteHosts()
 	v.Log.Infof("VCHub Destroyed")
 }
